requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
pillow==11.1.0
python-dotenv==1.0.1

# Postgres driver
psycopg2-binary==2.9.9

# Celery
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
